+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 -   Set EDC version to `0.0.1-20221006-SNAPSHOT`
--   Renamed `business-partner-validation` extension to `policy-validation` extension
+-   Renamed `business-partner-validation` extension to `access-policy-validation` extension
 -   Business Partner Number `Constraint` no longer supports the 'IN' constraint operator
 
 ### Changed